container/requirements.txt
requests==2.28.1
boto3==1.24.75
Pillow==9.0.1
sanic==22.9.1
ydb==2.10.0
pydantic==1.9.1
